Struct bejeweled::Game

pub struct Game {
    pub moves: i32,
    pub score: i32,
    pub grid: Vec<Vec<Color>>,
    pub start_time: u64,
    pub time_limit: bool,
}

A game is the collection of stats about the game and its current grid

Fields
moves: i32

Count of moves so far

score: i32

The current score of the game

grid: Vec<Vec<Color>>

8x8 grid of colors

start_time: u64

Start time, unix timestamp in ns

time_limit: bool

Defines whether or not the game has a time-limit

Methods

impl Game

fn new(time_limit: bool) -> Game

Create a new game, creates a random board then ensures that it's in a valid state before returning

fn print_board(&self)

A little helper function to pretty-print the board

fn make_move(&mut self, mov: &Move) -> bool

The top level function to make a move on the board, only executed if the move is valid, and "undone" if it doesn't score. Will return true if a move was made or false if the move could not be made, either because it was invalid or because the time is out

fn execute_move(&mut self, mov: &Move)

Swap the places of two colors on the grid, this could be optimized to remove cloning

fn pieces_to_remove(&self) -> Vec<Cell>

Gets all the pieces to remove from the board both row-wise and column-wise